Added constellation saving and loading; Including necessary data api;
Constellations are saved in this format:
One 32 bit number, giving star count and name: 8 bits star count, 4x6
bits letters of name
Following that are star_count number of bytes representing the stars in
the constellation. The lower 7 bits of the bytes give the index of the
star in the stars array. The first bit indicates whether this star
repeats in the sequence. I.e. if a constellation has
star1,star4,star4,star5, the serialised data would be: (repeat 0, index
1), (repeat 1, index 4), (repeat 0, index 5). This is useful both to
save space and also to make star info align along 8-bit boundries.