Fixed bug in connection deletion
Using the actual star instances to remove them from the constellation
caused problems if the same star was used in a previous connection of
the constellation. Then, not the intended connection would be deleted,
but whatever part of a connection that used that star before.
Deletion now uses specific indices and a function to do that deletion.